requirements.txt
pillow
requests
base64
flask
python-dotenv